+++ src/lib/components/BigButton.svelte
@@ -8,7 +8,7 @@
 </script>
 
 <Button.Root
-	class="flex items-center rounded-lg font-{fontWeight} px-4 py-2 text-slate-100 bg-{color}-700 hover:bg-{color}-600
+	class="inline-flex items-center rounded-lg font-{fontWeight} px-4 py-2 text-slate-100 bg-{color}-700 hover:bg-{color}-600
         {outline && `border-${color}-600 bg-opacity-0 border`}
         disabled:bg-gray-600/80 disabled:hover:bg-gray-600/80 disabled:text-gray-200/80 disabled:cursor-not-allowed"
 	on:click