+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 ### Fixed
 
+- Re-added CLI (oops)
+- FINALLY fixed the "Install with Mod Manager" button on Thunderstore
 - Various issues related to config file names with non-UTF8 characters
 - Multiple config files from the same mod being displayed with the same name
 - Decrease config time by reading files in parallel